refactor: drop unused modules and inline classifier
