--- src/compare/__init__.py
@@ -1,3 +1,0 @@
-from .cache_diff import backup_results, update_cache, diff_market
-
-__all__ = ["backup_results", "update_cache", "diff_market"]
