+++ src/meta/__init__.py
@@ -1,5 +1,3 @@
 """Meta utilities."""
 
-from .field_classifier import classify_fields
-
-__all__ = ["classify_fields"]
+__all__: list[str] = []
