--- src/spec/__init__.py
@@ -1,15 +1,0 @@
-"""Helpers for OpenAPI spec generation."""
-
-from .generator import (
-    generate_spec_for_all_markets,
-    generate_spec_for_market,
-    detect_all_markets,
-)
-from .bundler import bundle_all_specs
-
-__all__ = [
-    "generate_spec_for_all_markets",
-    "generate_spec_for_market",
-    "detect_all_markets",
-    "bundle_all_specs",
-]
